+++ R/cron.R
@@ -20,7 +20,8 @@
 #' \itemize{
 #'   \item a single integer value, like \code{1}
 #'   \item consecutive-integer vectors, like \code{1:3}
-#'   \item irregularly-spaced integer vectors, like \code{c(2, 3, 5)}
+#'   \item nonconsecutive, irregularly-spaced integer vectors, like
+#'       \code{c(2, 3, 5)}
 #'   \item regularly-spaced integer sequences with specified start and end
 #'   values, like \code{seq(3, 59, 15)} (useful for specifying sequences within
 #'   the full time period,'every 15 minutes of the hour starting from minute 3',
@@ -31,14 +32,15 @@
 #'     discretion).
 #' @export
 #'
-#' @examples \dontrun{r2cron(
-#'     minutes = seq(0, 59, 20),
-#'     hours = 15:17,  # 24-hr clock
-#'     days_month = 1,
-#'     months = c(4, 10, 11),
-#'     days_week = c(1, 7),  # Sunday is '1'
-#'     clip = FALSE
-#'     )}
+#' @examples \dontrun{
+#' r2cron(
+#'   minutes = seq(0, 59, 20),
+#'   hours = 15:17,  # 24-hr clock
+#'   days_month = 1,
+#'   months = c(4, 10, 11),
+#'   days_week = c(1, 7),  # Sunday is '1'
+#'   clip = FALSE
+#' )}
 #'
 r2cron <- function(minutes = 0L:59L, hours  = 0L:23L,
                    days_month  = 1L:31L, months = 1L:12L, days_week  = 1L:7L,
@@ -97,14 +99,16 @@ r2cron <- function(minutes = 0L:59L, hours  = 0L:23L,
 #' equivalent sentence in English. Can take the output from \code{\link{r2cron}}
 #' for example. (Under development.)
 #'
-#' @param cron Character. A valid cron string composed of five cron expressions
-#'     separated by spaces).
+#' @param cron Character. A valid cron string, i.e. one cron expression for each
+#'     of the five time period slots (minutes, hours, days of the month, months,
+#'     days of the week), separated by spaces.
 #'
 #' @return Character. An English sentence interpretation of the input cron
 #'     expression.
 #' @export
 #'
-#' @examples \dontrun{cron2eng("1,2,5 2-3 * 1/3 5")}
+#' @examples \dontrun{
+#' cron2eng("1,2,5 2-3 * 1/3 5")}
 #'
 cron2eng <- function(cron = "* * * * *") {
 
@@ -133,14 +137,17 @@ cron2eng <- function(cron = "* * * * *") {
 
       if (p_list[[period]] == "*") {
 
+        # Every valid unit of the time period
         p_list[[period]] <- paste("every", period)
 
       } else if (stringr::str_detect(p_list[[period]], "^\\d{1,2}$")) {
 
+        # Single integer value
         p_list[[period]] <- paste(period, p_list[[period]])
 
       } else if (stringr::str_detect(p_list[[period]], "^\\d{1,2}-\\d{1,2}$")) {
 
+        # Consecutive units with non-min start and non-max stop value
         n_split <- strsplit(p_list[[period]], "-")
         p_list[[period]] <- paste(
           period, n_split[[1]][1], "to", n_split[[1]][2]
@@ -148,11 +155,13 @@ cron2eng <- function(cron = "* * * * *") {
 
       } else if (stringr::str_detect(p_list[[period]], "\\d{1,2},")) {
 
+        # Nonconsecutive, irregularly spaced integers
         n_spaced <- stringr::str_replace_all(p_list[[period]], ",", ", ")
         p_list[[period]] <- paste(period, n_spaced)
 
       } else if (stringr::str_detect(p_list[[period]], "\\d{1,2}/\\d{1,2}")) {
 
+        # regularly-spaced integer sequences with specified start and end
         n_split <- strsplit(p_list[[period]], "/")
         p_list[[period]] <- paste(
           "every", n_split[[1]][2], period,
@@ -163,6 +172,7 @@ cron2eng <- function(cron = "* * * * *") {
 
     }
 
+    # Collate sentences for each time period
     paste(p_list, collapse = "; ")
 
   }