+++ tests/testthat/test-cron.R
@@ -9,7 +9,9 @@ test_that("minutes are viable", {
  expect_error(r2cron(minutes = 1.2, clip = FALSE))
  expect_error(r2cron(minutes = 60, clip = FALSE))
  expect_equal(r2cron(minutes = 1:3, clip = FALSE), "1-3 * * * *")
+ expect_equal(r2cron(minutes = 1:2, clip = FALSE), "1-2 * * * *")  # maybe 1,2
  expect_equal(r2cron(minutes = seq(0, 59, 15), clip = FALSE), "0/15 * * * *")
+ expect_equal(r2cron(minutes = c(1, 6), clip = FALSE), "1,6 * * * *")
  expect_equal(r2cron(minutes = c(1, 6, 20), clip = FALSE), "1,6,20 * * * *")
 })
 
@@ -20,7 +22,9 @@ test_that("hours are viable", {
  expect_error(r2cron(hours = 1.2, clip = FALSE))
  expect_error(r2cron(hours = 24, clip = FALSE))
  expect_equal(r2cron(hours = 1:3, clip = FALSE), "* 1-3 * * *")
+ expect_equal(r2cron(hours = 1:2, clip = FALSE), "* 1-2 * * *")
  expect_equal(r2cron(hours = seq(0, 23, 3), clip = FALSE), "* 0/3 * * *")
+ expect_equal(r2cron(hours = c(1, 6), clip = FALSE), "* 1,6 * * *")
  expect_equal(r2cron(hours = c(1, 6, 20), clip = FALSE), "* 1,6,20 * * *")
 })
 
@@ -31,7 +35,9 @@ test_that("days_month are viable", {
  expect_error(r2cron(days_month = 1.2, clip = FALSE))
  expect_error(r2cron(days_month = 32, clip = FALSE))
  expect_equal(r2cron(days_month = 1:3, clip = FALSE), "* * 1-3 * *")
+ expect_equal(r2cron(days_month = 1:2, clip = FALSE), "* * 1-2 * *")
  expect_equal(r2cron(days_month = seq(1, 31, 3), clip = FALSE), "* * 1/3 * *")
+ expect_equal(r2cron(days_month = c(1, 6), clip = FALSE), "* * 1,6 * *")
  expect_equal(r2cron(days_month = c(1, 6, 20), clip = FALSE), "* * 1,6,20 * *")
 })
 
@@ -42,7 +48,9 @@ test_that("months are viable", {
  expect_error(r2cron(months = 1.2, clip = FALSE))
  expect_error(r2cron(months = 13, clip = FALSE))
  expect_equal(r2cron(months = 1:3, clip = FALSE), "* * * 1-3 *")
+ expect_equal(r2cron(months = 1:2, clip = FALSE), "* * * 1-2 *")
  expect_equal(r2cron(months = seq(1, 12, 3), clip = FALSE), "* * * 1/3 *")
+ expect_equal(r2cron(months = c(1, 6), clip = FALSE), "* * * 1,6 *")
  expect_equal(r2cron(months = c(1, 6, 7), clip = FALSE), "* * * 1,6,7 *")
 })
 
@@ -53,11 +61,14 @@ test_that("days_week are viable", {
  expect_error(r2cron(days_week = 1.2, clip = FALSE))
  expect_error(r2cron(days_week = 8, clip = FALSE))
  expect_equal(r2cron(days_week = 2:4, clip = FALSE), "* * * * 1-3")
+ expect_equal(r2cron(days_week = 2:3, clip = FALSE), "* * * * 1-2")
  expect_equal(r2cron(days_week = seq(1, 7, 2), clip = FALSE), "* * * * 0/2")
+ expect_equal(r2cron(days_week = c(1, 3), clip = FALSE), "* * * * 0,2")
+ expect_equal(r2cron(days_week = c(1, 7), clip = FALSE), "* * * * 0,6")  # outputs 0/6, why?
  expect_equal(r2cron(days_week = c(1, 3, 4), clip = FALSE), "* * * * 0,2,3")
 })
 
-test_that("out-of-range values are warned", {
+test_that("out-of-range day-month values are warned", {
  expect_warning(r2cron(days_month = 29, months = 2, clip = FALSE))
  expect_warning(r2cron(days_month = 30, months = 2, clip = FALSE))
  expect_warning(r2cron(days_month = 31, months = 2, clip = FALSE))